+++ freddy-cruder-spring-data-jpa/src/main/java/com/peluware/freddy/cruder/springframework/jpa/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.peluware.freddy.cruder.springframework.jpa;
+
+import org.jspecify.annotations.NullMarked;